projects/bvs/da-profissionais-liberais-development/requirements.txt
aiodns>=3.1.1
aiohttp>=3.8.6
apache-airflow>=2.6.3
fastparquet>=0.8.1
google-api-core>=2.15.0
google-api-python-client>=2.113.0
google-auth>=2.26.2
google-auth-httplib2>=0.2.0
google-cloud-bigtable>=2.22.0
google-cloud-core>=2.4.1
google-cloud-dataproc>=5.8.0
google-cloud-storage>=2.14.0
googleapis-common-protos>=1.62.0
nest-asyncio>=1.5.8
pandas>=1.3.5
pyspark>=3.4.2
requests>=2.31.0
selenium>=4.11.2
webdriver-manager>=4.0.1
selenium-wire>=5.1.0
